+++ clouds/bigquery/libraries/javascript/src/random.js
@@ -1,0 +1,5 @@
+import { bbox } from '@turf/turf';
+
+export default {
+    bbox
+};