+++ clouds/bigquery/libraries/javascript/test/clustering.test.js
@@ -1,0 +1,7 @@
+const lib = require('../build/index');
+
+test('clustering library defined', () => {
+    expect(lib.clustering.featureCollection).toBeDefined();
+    expect(lib.clustering.feature).toBeDefined();
+    expect(lib.clustering.clustersKmeans).toBeDefined();
+});