+++ clouds/bigquery/libraries/javascript/test/random.test.js
@@ -1,0 +1,5 @@
+const lib = require('../build/index');
+
+test('random library defined', () => {
+    expect(lib.random.bbox).toBeDefined();
+});