+++ clouds/snowflake/libraries/javascript/test/random.test.js
@@ -1,0 +1,5 @@
+const randomLib = require('../build/random');
+
+test('random library defined', () => {
+    expect(randomLib.bbox).toBeDefined();
+});